asaas_split_app/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Werkzeug==2.2.3
requests==2.32.3
python-dotenv==1.0.1
